+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Send report with the scheduler actions periodically
 - Output to a log file using logrus library
 - Log verbosity defined in configuration
+- BoltDB storage for tracking backups locally
 
 ### Fixed
 - Add sample configuration file to deb and txz packages